requirements.txt
scikit-learn
spacy
sentence-transformers
pdfplumber
uvicorn
fastapi
pydantic
python-dotenv
